Add constructor, options, and visitor tracking.
The New(...) constructor will return a new Limiter and would be called
once per route setup, or once globally, or however else the implementor
wishes to apply rate limiting.
Added properties to the Limiter struct to store configuration choices,
as well as a placeholder in-memory map that will keep track of visitors
until the Redis implementation is complete, and a mutex for locking
said map and making it thread-safe.
The visitors are identified by an arbitrary string ID and the timestamp
of each visit is stored, so this provides a foundation upon which to
build more complex logic around sliding windows.